dep cleanup

+++ deps/vrx.py
@@ -6,60 +6,64 @@
 # see http://www.gnu.org/licenses/gpl-2.0.html
 ###############################################################################
 
-VERSION ="master"
+VERSION = "master"
 
 import os, tarfile
-from ork import dep, host, path, git, cmake, make
+from yarl import URL
+from ork import dep, host, path, cmake, git, make
 from ork.deco import Deco
 from ork.wget import wget
 from ork.command import Command
-from ork.cmake import context
 
 deco = Deco()
 
 ###############################################################################
 
-class assimp(dep.Provider):
+class vrx(dep.Provider):
 
   def __init__(self,options=None): ############################################
 
-    parclass = super(assimp,self)
+    parclass = super(vrx,self)
     parclass.__init__(options=options)
-
-    self.source_dest = path.builds()/"assimp"
-    self.build_dest = path.builds()/"assimp"/".build"
-    self.manifest = path.manifests()/"assimp"
-
+    #print(options)
+    self.source_dest = path.builds()/"vrx"
+    self.build_dest = path.builds()/"vrx"/".build"
+    self.manifest = path.manifests()/"vrx"
     self.OK = self.manifest.exists()
 
   def __str__(self): ##########################################################
 
-    return "Assimp (homebrew)"
+    return "VRX (%s-source)" % VERSION
 
   def build(self): ##########################################################
 
-    git.Clone("https://github.com/assimp/assimp",self.source_dest,VERSION)
+    assimp = dep.require("assimp")
+    glfw = dep.require("glfw")
+    utpp = dep.require("unittestpp")
 
-    os.system("rm -rf %s"%self.build_dest)
-    os.mkdir(self.build_dest)
-    os.chdir(self.build_dest)
-    cmake_ctx = cmake.context("..")
-    cmake_ctx.exec()
-    return (make.exec("install")==0)
+    if self.incremental():
+        os.chdir(self.build_dest)
+    else:
+        git.Clone("https://github.com/tweakoz/vrx",self.source_dest,VERSION)
+        os.system("rm -rf %s"%self.build_dest)
+        os.mkdir(self.build_dest)
+        os.chdir(self.build_dest)
+        cmake_ctx = cmake.context("..")
+        cmake_ctx.exec()
+
+    rval = (make.exec("install")==0)
+    return rval
 
   def linkenv(self): ##########################################################
+    LIBS = ["ork_vrx"]
     return {
-        "LIBS": ["assimp"]
+        "LIBS": LIBS,
+        "LFLAGS": ["-l%s"%item for item in LIBS]
     }
-
   def provide(self): ##########################################################
 
     if self.should_build():
-
       self.OK = self.build()
       if self.OK:
-        self.manifest.touch()
-
-    print(self.OK)
-
+          self.manifest.touch()
     return self.OK
